Refactor codebase by splitting  crawling functions into separate files
