+++ src/gcp_scanner/gcp_crawlers/__init__.py
@@ -1,0 +1,15 @@
+"""The module to query GCP resources via RestAPI.
+
+"""
+
+from .compute_crawler import *
+from .db_crawler import *
+from .gke_crawler import *
+from .mq_crawler import *
+from .network_crawler import *
+from .project_crawler import *
+from .security_crawler import *
+from .serverless_crawler import *
+from .service_account_crawler import *
+from .sourcerepo_crawler import *
+from .storage_crawler import *
